[swarming] Build and store bot archive in the datastore.
Bot archives are only ~1.5 MB in size. I decided it is OK to use
datastore for storing them to avoid adding a dependency on some
other storage system like GCS. It's not worth the effort.

In the python backend version they are essentially stored in
memcache. But we don't have it in the Go version. Instead for
actually serving files, we'll cache archives in the local
process memory (again, this is just ~1.5 MB which is
insignificant for our F4 GAE instances that have 1536 MB RAM
limit; we also have an option of using F4_1G instances with
3072 MB of RAM).

R=chanli@chromium.org
BUG=b/355013257

Change-Id: I4a9a7aedd92e3c526b6074a1f75539681f076abf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5805747
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>