ui: less flickering for monitoring page
After the previous CL, flickering was reduced, but not eliminated.
Another source of flickering on the page is that when the alert query
succeeds, it changes the key of the extended alerts query, putting it
back into a loading state, and thus eliminating all components on the
page (to be re-created after extended a;erts query succeeds).

To avoid this, set keepPreviousData so that the data is held stable and
there is no loading state transition during the new fetch.  This seems
to work well in my testing, I didn't see a flicker in ~20 minutes.

Change-Id: Icbe7ceb8696a9cefee2a6a1b4f34def8b725dc66
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5820223
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Auto-Submit: Matthew Warton <mwarton@google.com>
Commit-Queue: Matthew Warton <mwarton@google.com>