[server] Bind listening sockets ASAP.
This matters for startup probes. If a socket is already bound when
the server is starting, the connection to it from the startup probe
will wait until the server run loop is running, and the probe will
succeed as soon as this happens (usually takes 1-2 sec on GCP).

If we bind the socket too late, the first startup probe instead will
fail with ECONNREFUSED, which will trigger a retry after some
constant timeout (which is usually 5-10s). This results in few
seconds of unnecessary delay.

R=iannucci@chromium.org

Change-Id: Ibb9f62861ca139e553d73e28facccbf18d7fdd2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5826969
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>