[server/tq] Report queue name in monitoring metrics.
We are adding a feature that would allow a single task class
to run on many different queues (for things like sharding).
Reporting to monitoring just the task class alone is no
longer sufficient to see the full picture of what is going on.

R=weiweilin@google.com

Change-Id: I741be8a3daa68ccdb9ea24b1487f5db417f55bdf
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5826975
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>