[server/tq] Allow a single task class to run on many queues.
A QueuePicker callback can be used to decide a queue name for
each individual instance of a task. It can be used to distribute
tasks across multiple queues, for example to spread the load or
to use different queues for tasks with different priorities.

R=weiweilin@google.com

Change-Id: Id4686b9461d17bbea82cc24ae88b8cfbf4c78746
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5838079
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>