[swarming] Use server/pubsub module instead of a hand-rolled handler.
This is a new server module that allows exposing PubSub push
handlers. Its advantage over the handrolled handler is that it
also reports some monitoring metrics.

This CL exposes a new route "/internal/pubsub/rbe-scheduler", but
also keeps the old one. The subscription will be changed to push
to the new route, after which the old code can be deleted.

R=chanli@chromium.org

Change-Id: Ic09384086d1f397c8f96dfbfc399c0f34131acd2
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5838659
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>