[swarming] Implement GetDetails RPC.
It was blocked on inability to calculate bot_version from
the Go side, but it is no longer a problem.

Most of this CL is mocking of the bot CIPD package, which is
a bit of an overkill for this CL specifically, but will be
reused in tests for bot RPC handlers.

R=chanli@chromium.org

Change-Id: Ifaee541ed2f8e0e171df9b3dc344cb73a463ae61
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5841539
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>