[swarming] Implement GetToken RPC.
It uses the secret stored in the datastore (and a bunch of
ugly glue code) for now for compatibility with Python code.

Once all relevant code is in Go, this will be updated to use
a more modern approach based on Tink and Google Secret Manager.

R=chanli@chromium.org
BUG=b/355013257

Change-Id: Idb0901e81756f7d8c6ba7af4d5483e394dce36a0
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5846408
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>