[swarming] Improve config caching logic.
There's a problem when using config cached in the local
process memory when a bot hits two different backend
processes.

The following can happen (and happens for real in
the Python implementation):
  1. The bot process calls the backend process S1 and
     gets a response based on the config revision R2.
  2. The same bot process calls another backend process S2,
     and gets a previous revision R1, since S2 has a slightly
     stale cache.
  3. The bot process calls the backend process S3 and
     gets R2 again.

This results in the "flapping" of the bot between two
configs (most noticeable when the config being changed
is the bot hooks script: this causes the bot to restart
multiple times in a row until the backend cache is
fully updated across all processes).

An individual bot process in isolation should not see
configs going back in time. It should see the config going
only forward in time.

One way to accomplish that is to stop using the config
cache and instead always hit the datastore to get the
up-to-date config. But this is problematic for several
reasons:
  1. The datastore entity with the config becomes a bright
     hot spot (thousands QPS fetching it), causing reliability
     problems or complexity (if sharding it).
  2. Unnecessary datastore costs (though minor comparatively
     speaking, but still avoidable).
  3. Extra request processing latency.

Instead we are going to make the bot process track what
is the last config revision it has seen, and send it with
every request to the server. If the bot claims it saw
R2, but the local cache on the backend has R1, then the
local cache is most definitely stale and we should fetch
the latest config from the datastore. But if the backend
has R2 already cached, then its fine to keep using it
(even if there's R3 already in the datastore: it doesn't
matter, neither participant has seen it yet).

In the steady state (when the config is not changing),
there will be no datastore fetches at all (other than
occasional periodic check for new configs).

This is a primitive variant of
https://en.wikipedia.org/wiki/Vector_clock

Note that we can get away with it because we don't care if
two **different** bots see different versions of configs.
We only care about the timeline of a single bot.

This CL only implements the helper for fetching the latest
config when necessary (it will be used when implementing
"/bot_code"). The rest of the described protocol will be
implemented later, when migrating bot API handlers to Go.

R=chanli@chromium.org
BUG=b/355013257

Change-Id: I617b02d8df8aa8a2c6b1aac58936b6837fa8b866
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5847768
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>