[luciexe/build] Refactor properties into their own package.
This package provides a much nicer type-safe interface for
registering, reading and manipulating protos. After this
implementation lands with tests, the main build library will be
refactored in terms of this.

Because this is also separate from the main `build` library, it also
means that this can be more easily used for programs operating in
'no-op' mode, as well as in tests.

R=azrahman, chanli, fancl

Change-Id: I10def23feaae65872822dc4d3f4e8da22ba77beb
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5835879
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>