[swarming] Authenticate bot calls based on bots.cfg in dry run mode.
Existing "/swarming/api/v1/bot/rbe/session/*" endpoints currently
use information from the poll token to figure out how to
authenticate requests. These poll tokens are HMAC-protected blobs
of information that are prepared by the Python side and passed
through the bot to the Go side. The Python side constructs them
based on bots.cfg configs.

This was added to avoid blocking RBE migration on implementing
non-trivial config ingestion and parsing subsystem in Go.

We now have this subsystem and we should stop relying on poll
tokens. We'll forced to do so for "/bot_code" handler because
it is the first ever call from a bot to the backend (there's
no poll token there yet).

This CL hooks up the authentication code, but doesn't yet
use its results for real. It just compares the outcome of
the authentication (and some other information) to what's
in the poll token. Once we confirm there's no violations,
poll tokens can be retired.

R=chanli@chromium.org
BUG=b/355013257

Change-Id: I5efbe55cad5d0d0be62504d69878be7c362a6cce
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5855641
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>