[common/testing/truth/should] Make should.(Not)?Equal fail on time.Time
This is a very common error - Goconvey had the following traps:
  * ShouldEqual was secretly ShouldResemble
  * ShouldResemble secretly had special conditions for comparing
    timestamps (specifically using the .Equal method).

In `truth`, should.Equal strives to be the same as `==` in Go
- struct types (like time.Time) ARE comparable with `==`, except
that this comparison is incorrect and leads to confusing errors
because of the hidden monotonic time fields.

This CL special cases the time.Time type with an explanation of why
should.Equal is the wrong comparison type, and how to remedy.

This also fixes all of the existing incorrect uses of
should.Equal[time.Time] in the luci-go repo.

R=ddoman, gregorynisbet@google.com, yiwzhang

Bug: 319730573
Change-Id: I0cb98a57468fd6ed4a2657979155c6eb2ddbf930
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5856021
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>