[prpc] Introduce "X-Prpc-Max-Response-Size" request header.
It can be used by the client to tell the server the maximum
response size the client is willing to deserialize. The
server may use it to discard large responses without trying
to send them to the client, thus avoiding wasting everyone's
time.

For backward compatibility we'll need to keep using UNAVAILABLE
error for this condition. But this error usually triggers a retry.
To allow clients to skip retrying "response too large" errors in
particular (since they are very often are non-retriable), we attach
a new luci.prpc.ErrorDetails proto to google.rpc.Status.

Fix some formatting nits in the prpc protocol doc and add a section
about alternatives.

Also reduce DefaultMaxContentLength by 32KiB. It was set to 32MiB
exactly, which caused requests close to the limit to exceed GAE
thresholds (since there's some overhead on wrapping the proto
response into an HTTP response).

R=iannucci@chromium.org
BUG=b/361633387

Change-Id: Ie5b55bbebff27a7e49f96bd5d2ce7e2deb8a7973
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5860805
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>