[prpc] Fix unmarshalling of X-Prpc-Status-Details-Bin.
The spec says it should be serialized using the same encoding
as the response (based on the "Accept" header). And the server
does that. But the client always assumed it is wirepb encoding.

As a result, error details didn't really work when using JSON
response encoding.

R=iannucci@chromium.org
BUG=b/361633387

Change-Id: I293059b021bf196e642a73ad5f29c20b33ba70a4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5867729
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>