[server] Do not log panics due to ErrAbortHandler.
Its doc is:
  ErrAbortHandler is a sentinel panic value to abort a handler.
  While any panic from ServeHTTP aborts the response to the client,
  panicking with ErrAbortHandler also suppresses logging of a stack
  trace to the server's error log.

In particular we see this when using httputil.ReverseProxy to
proxy requests from Go to Py (as a traffic splitting strategy
when gracefully migrating from Py to Go). If a client disconnects,
the ReverseProxy panics with ErrAbortHandler. This ends up spamming
our server logs.

R=iannucci@chromium.org, chanli@chromium.org

Change-Id: I3c3eb6f77749f9dc982557ad09f1da03cd985a3c
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5867659
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>