[tqtesting] Fix race condition in tqtesting.Scheduler.
It was checking the stop conditions and dequeuing tasks
in two separate critical sections. As a result,
it was possible for a task specified in StopBeforeTask(...)
stop condition to actually be dequeued and executed (if it
happens to appear asynchronously after shouldStop, but before
tryDequeueTask).

Looks like this was somehow resulting in one of the tests
getting stuck (this test executes and submits tasks
asynchronously).

To avoid that, check the stop conditions and dequeue tasks
under a single critical section.

Note that this is all in the "local" Cloud Tasks emulation
used in tests, this should have 0 effect on any production
usage of server/tq.

R=iannucci@chromium.org
BUG=b/367495072

Change-Id: I8ff2471a3a3b241444ed94822d8ec039fc2c11d4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5868622
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>