[TreeStatus] Set project params
I noticed the top bar of the Tree Status page shows the tree ID instead of the project name, e.g. https://screenshot.googleplex.com/BDuiC3B4tcbgF6y . So when the use click on the tree name, it may go to an invalid project, if the tree name and the project name is different. The bug seems to be there for quite long.

So I pass in an optional project field so the project will be set correctly when a user goes to the TreeStatus from the sidebar.

Note: If the user access the tree directly (without the sidebar), the top bar still show the tree name. This can be fixed in later CL.

Demo: https://16782-9ec59f1-nqmtua-treestatus-set-correct-p-dot-luci-milo-dev.appspot.com/ui/tree-status/chromium?project=chromium


Bug:b/363155263
Change-Id: Ib0e5651409f8472fa61f6f91f32a23e21372563b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5868355
Auto-Submit: Tuan Nguyen <nqmtuan@google.com>
Reviewed-by: Weiwei Lin <weiweilin@google.com>
Commit-Queue: Weiwei Lin <weiweilin@google.com>