[prpc] Fix handling of reading errors in the override callback.
It is possible (and indeed happens in Swarming [1]) for the override
callback to "peek" at the request, see an error, but then ignore it
and let request processing happen anyway (with assumption that this
error will be caught again in the appropriate place).

This didn't work correctly, because req.Body was not replaced
on errors. As a result, the pRPC server (or the reverse HTTP
proxy in Swarming's case when redirecting to Py) essentially
just saw an empty request (because req.Body was exhausted
already) and failed with EOF error.

It is not a super critical bug though, since the request was
bad to begin with. It was just failing with a wrong error
(EOF instead of a more descriptive request deserialization
error).

Now the original request body is fully reproduced (including
any IO errors), even if the message is malformed.

[1]: https://chromium.googlesource.com/infra/luci/luci-go/+/b91a845/swarming/server/rpcs/migration.go#150

R=chanli@chromium.org

Change-Id: I6428084272a1f68efdf7b788e83655fc166bab27
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5889396
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>