[prpc] More attempts to deflake e2e test.
Adjust timeouts to be less tight. It seems some tests
are failing when requests that are supposed to be "fast"
are actually slow (due to the CPU being overloaded with
other tests). Such tests hit timeouts unexpectedly.

Note that this test specifically does not mock time, since it
specifically tests real world behaviors involving real
clocks, real sockets, etc.

The test is pretty slow now (10s), since it sits and waits for
these longer timeouts to expire...

Another change is making the test service return an error if
the request context is canceled. There's a benign expected race
when an RPC implementation returns a result despite its context
being canceled. This result is accepted and returned to the
client (why drop it if we have it?). But in this particular test
this race causes some flakiness, so make the service implementation
explicitly return an error if the context is canceled.

R=iannucci@chromium.org
BUG=b/325867011

Change-Id: I16994127b3172574a74287c156ae4915df2082e6
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5893286
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>