[dispatcher] Fix subtle race in coordinator.
Prior to this CL, `run()` can fully terminate, thus closing
`state.resultCh` while some goroutines which are about-to-push into
this channel are still scheduled. When those scheduled goroutines
eventually wake up, they push into a now-closed channel and panic,
crashing the program.

This CL changes the behavior of MinQPS so that it correctly
interacts with the buffer state (and thus the state of `run`) when
it does fire.

See comments in-CL for detailed explanation.

R=chanli, dmitshur@google.com, vadimsh

Bug: go.dev/issue/69276
Change-Id: I78f225d3ee493a39fcbebfc051388dcc536f663e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5896786
Commit-Queue: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>