[spantest] Refactor Spanner testing implementation.
Use new Emulator wrapper that uses OS-assigned ports.
Make it super explicit when a REAL SPANNER is used.
De-copy-paste findInitScript.
Stop messing with gcloud config, it doesn't do anything.
Don't drop DBs on emulator, it is a waste of time.

ResultDB tests have a partial copy-pasta of spantest package,
and they use real spanner DB by default. I'm not sure if this
is intentional or not, so this is preserved. Since this is likely
not something we want to support, remove implicit fallback on
the real spanner from "spantest" package. The decision to use
a real database now should be made explicitly by the test author.

R=iannucci@chromium.org

Change-Id: I8f1a16b4d78500becdeccc23f74633f03d7765cc
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5908632
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>