[buildbucket] Rewrite the rest of the tests to ftt/truth.
This required a bit of manual work to adjust how fakeCfgClient
works. For some reason, even though the client was installed in the
context, it relied on the manipulation of global variables to change
the value exposed by the client back to the code under test.

This changes the client so that it internally holds all
configuration data, which decouples all the test cases from each
other.

R=chanli@chromium.org

Change-Id: I96a2720914c6590f9da39a87f3736b0918506ef3
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5922042
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>