[server] Stop using deprecated otelgrpc's server interceptors.
They have been replaced with a stats.Handler. This is not a drop-in
replacement for us, since stats.Handlers are executed before all
interceptors, and in particular before the interceptor that
initializes per-request context.

The per-request context now needs to be initialized in a
stats.Handler as well (to run before otelgrpc uses it). We care about
that because OTEL sampler we use (which is a process-global thing)
looks at the context given to it and makes different decisions for
per-RPC context vs some other random context (e.g. background
contexts started by background goroutines).

R=iannucci@chromium.org

Change-Id: I683961ea7c7dafc3907313accf60de3a20fa4811
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5938494
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>