[aip160] Extract filter_parser into aip160 library in common/data
As discussed in the LUCI Users chat room, I want to start supporting AIP160
syntax more generally across more of our projects. LUCI analysis already has
an implementation of a parser for this that is pretty complete, so let's extract it
to a common area for use in future projects.

For now, we're only promoting the filter parsing logic. Future CLs will add functionality
related to serialization and define a common proto for AIP160 ASTs.

BUG=369104261
TEST=make test

Change-Id: Icc43ba3f77664feaaeaad4959bb33270aa8a74a4
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5958388
Reviewed-by: Robbie Iannucci <iannucci@chromium.org>
Reviewed-by: Matthew Warton <mwarton@google.com>
Commit-Queue: Gregory Nisbet <gregorynisbet@google.com>