[tsmon] Remove SetFixedResetTime.
It is hard to use correctly. Removing it also simplifies API a bit.

It also opens a possibility for implementing ProdX Mon API contract
correctly (one reset timestamp **per target**, not per metric). But
this hasn't been done in this CL.

This also fixes a deficiency added in
https://chromium.googlesource.com/infra/luci/luci-go/+/bfbb3154de49
It called inMemoryStore.Now(...) quite often, since for most
metrics m.fixedResetTime was zero. This is bad, since Now(...) hits
a global mutex (though tsmon is already full of them, so it didn't
really make a difference, but not hitting a mutex all the time is
still nicer).

R=ddoman@chromium.org

Change-Id: I1ba5adfee0b1c5766d1b1fdb0ff4176cdfe5f722
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5962817
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>