[buildbucket] Speculative fix for handling errors when batch scheduling.
This changes the error handling code to make sure that every item in
the returned batch is either a Build or an error, with an error
taking precedence.

I'm not certain that this will actually fix the issue, but the code
location in question seemed like it could have allowed a non-nil,
but empty, build to override an error in the response.

R=chanli

Bug: 376505379
Change-Id: I81944e5c689c228bb65d7cd21c1d3398074bc10b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5980790
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>
Auto-Submit: Robbie Iannucci <iannucci@chromium.org>