[prpc] Refactor serialization implementation.
The primary goal is to concentrate serialization implementation
in a single place, to be able to change it later via simple
tweaks.

The secondary goal is to remove protobuf v1 messages from the
public interface of pRPC packages and tests (and keep it
contained to a single file).

This introduces a notion of a codec. A codec takes protobuf v2
message and encodes it using some rules into a byte blob. This blob
has one of the 3 formats: Binary, Text or JSON.

For a given format, there are multiple codecs that can work with it:
  * Binary: codecWireV1 and codecWireV2.
  * JSON: codecJSONV1, codecJSONV1WithHack, codecJSONV2.
  * Text: codecTextV1, codecTextV2.

Where v1 stands for "using protobuf v1 rules" and v2 stands for
"using protobuf v2 rules" (because they are different).

The server and the client use their configuration to decide what
codec to use for a particular format. The current logic 100%
matches the previous implementation, i.e. this should be a pure
refactoring that doesn't change any observable behavior.

R=iannucci@chromium.org, gregorynisbet@google.com
BUG=b/376137855

Change-Id: I6ca58963de8764e6ccee96a8ce2de3eecee87132
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5974555
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>