[auth-service] Increase AuthGroups cache refresh frequency
This should help with the ListGroups latency. Previous refresh
was every 5 min, but that is also how frequently the
groups-push-cron runs. Having a more frequent refresh of the
groups cache should hopefully reduce the instances in which the
cache is outdated and all groups need to be fetched when a
ListGroups call comes in.

Change-Id: Ifd705db57947b43f57ac6765f4caf06e88e51b95
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5981226
Commit-Queue: Rachael Newitt <renewitt@google.com>
Reviewed-by: Rachael Newitt <renewitt@google.com>
Auto-Submit: Anne Redulla <aredulla@google.com>