[Fleet Console] Create a sandbox page for adding code demos with sample direct UFS integration
Demo: https://17105-0f99334-tainted-zhangtiff-ping-dot-luci-milo-dev.appspot.com/ui/fleet/labs/sandbox (you'll need to login to see any interesting data)

This is test page that's not meant to be user-facing (but also isn't meant to be secret) - it's meant to be a freeform place where we can commit code meant to test things.

ie: I added code here that demonstrates that we can fetch data from UFS on the client. This is meant to be a proof of concept since we don't expect the client to directly fetch data from UFS for most cases, but including this sample here for reference.

Bug:358095025
Change-Id: Ie609dd212aac8e0b34af34df5e26608852218f2e
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5986997
Commit-Queue: Tiffany Zhang <zhangtiff@chromium.org>
Reviewed-by: Pietro Scuttari <pietroscutta@google.com>