[luci_machine_tokend] Add `-token-file-copy` flag.
It can be used to instruct luci_machine_tokend to put a copy of
the token file into multiple locations in addition to `-token-file`:

```
luci_machine_tokend \
    ...
    -token-file /path/to/primary/copy \
    -token-file-copy /path/to/copy1 \
    -token-file-copy /path/to/copy2
```

Note that `-token-file` is the special "primary" copy of the token
file, since we read only this file to know when the token expires
(thus it is still a separate flag).

Also add some tests.

R=chanli@chromium.org
BUG=b/372434320

Change-Id: I7295f2bd11ed9de55e40424f8e79a9b47d9ba22b
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/5995734
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>