[auth-service] Return general access denied error
The caller details and relevant groups can be logged internally instead
of telling the caller exactly which groups are required for the
endpoint.

Not that much of a stretch to find out which groups because this is
open-source code anyway, but there's no need to expose the groups
to web crawlers and whatnot.

Change-Id: Ib94bfc456bf744f113616143a4c8713111dcd540
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6026735
Reviewed-by: Joey Scarr <jsca@google.com>
Reviewed-by: Rachael Newitt <renewitt@google.com>
Commit-Queue: Anne Redulla <aredulla@google.com>