[server] Slightly improve experience of using ID tokens locally.
Some OAuth refresh token is required to generate an ID token when
using developer's credentials. The only way to get a refresh token
is to go through an interactive login flow in "luci-auth login".
Each such login creates a refresh token associated with some set of
OAuth scopes. By default this set is ["userinfo.email"].

TokenGenerator used to omit `auth.Options.Scopes` when generating
ID tokens. As a result, the default refresh token for
["userinfo.email"] was used.

This is bad because there's no guarantee it is present. When it is
not present, the token generation fails with "interactive login
required" error.

But there **is** as set of scopes that ~= guaranteed to have
a refresh token for them, because server.Server explicitly checks
this when it starts locally: it is the set of scopes passed as
part of the server's ClientAuth configuration (defaulting to
auth.CloudOAuthScopes). We should use it when making ID tokens.

Also move population of opts.ClientAuth.Scopes defaults to be
slightly later (in case the caller doesn't call OptionsFromEnv)
and actually check for presence of opts.ClientAuth.Scopes
specifically when starting, not just auth.CloudOAuthScopes. This
theoretically matters if the caller supplies their own custom set
of scopes for the local server for some reason (I'm not aware of
this happening anywhere though).

R=gregorynisbet@google.com
TEST=manually running a server that makes calls using ID tokens

Change-Id: I99a629117be2e9f807d7104a9b8ca9521634e404
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6032928
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>