[swarming] Switch to using session tokens for authentication.
To allow rolling back, also keep generating old-format session
token as long as the bot is sending necessary inputs. Once bots
are updated not to use old tokens, this can be finally removed.

Remove fakebot that was used long time ago to test stuff locally.
It is hard to keep it in working order. It may be resurrected in
some form later after all migrations are done.

R=chanli@chromium.org
BUG=b/355013282

Change-Id: I65d65688f5aa2aaa8f5f862043351e9b4cac6ca8
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6026773
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>
Reviewed-by: Chan Li <chanli@chromium.org>