[prpc] Recognize HTTP 401 and 403 errors from Cloud Run.
They can happen if the Cloud Run frontend is configured to
authenticate requests before passing them to the backend.
If a request fails, Cloud Run replies with HTTP 401 or
HTTP 403 status codes and a simple HTML page with the error
message.

This trips the pRPC client since it expects to see
X-Prpc-Grpc-Code response header with gRPC status code. Such
responses result in Internal errors (and useless retries).

Instead recognize HTTP 401 and 403 errors specifically
and convert them into the corresponding gRPC status errors.
This was already used for >=500 statuses (for similar reasons),
this CL extends this mechanism to cover other HTTP codes that
can potentially originate from proxies in front of the pRPC
server. Note that statuses as HTTP 400 are still not covered,
since proxies are not expected to return them.

R=gregorynisbet@google.com

Change-Id: I1d15872beb5e45e3176791bbe4a591acb545cb1f
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6036408
Reviewed-by: Gregory Nisbet <gregorynisbet@google.com>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>