[swarming] Implement BotVisitor interface for update_named_caches.
This is the base implementation of update_named_caches to satisfy the BotVisitor interface.

Unfortunately, the current json representation of 'named_cache' cannot be defined with a go struct. To bypass this limitation we use json.Decoder instead of json.Unmarshal. This coupled with Decoder.UseNumber, allows us to extract the size without losing precision.

BUG:b/355013356
Change-Id: I14dd2fa74d88d198ef7791ec9ee14255c05df25d
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6011278
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: JB Tamas-Leloup <jtamasleloup@google.com>