findings: prepend SeverityLevel enum value with SEVERITY_LEVEL
This is to follow the best practice described in:
https://protobuf.dev/programming-guides/style/#enums
Also change the first value to be UNSPECIFIED according to the
best practice.

The trigger of the change is that I find myself writing code
like `Finding.UNSET` today which leaves me clueless on what this
exactly means.

Change-Id: I056a6444bafef3b38f5c1b87b00772be5808c9a5
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6042363
Reviewed-by: Scott Lee <ddoman@chromium.org>
Commit-Queue: Yiwei Zhang <yiwzhang@google.com>