[cas-viewer] Fix how it creates gRPC clients.
It used to use context.Background() as the root context for
gRPC clients and its per-RPC credentials. This worked only
due to some fallback on per-request context that the auth
library is doing. But it is fragile and at some point it had
some changes that resulted in following error happening
in this situation:

> transport: per-RPC creds failed due to error: a transport or
credentials provider created within a context of one user request
is used within another user request, this is dangerous

This CL makes the client use the root server context that is
initialized to make gRPC auth working as expected.

Plus some style nits.

R=chanli@chromium.org

Change-Id: Ib922c08dab3c6dbfc23b9e5e2ef3b86a3db06d96
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6065160
Reviewed-by: Chan Li <chanli@chromium.org>
Commit-Queue: Vadim Shtayura <vadimsh@chromium.org>