[Buildbucket] Max Current Build: handle popping a pending build
In the case that a pending build is cancelled, we should clean it out from the pending queue.

Also add a build status check in backend task creation, given that there may be a gap between a build is scheduled and sent to backend.

Bug: 379656381
Change-Id: I462cd5a0ca0ac6dba9e2fbfce98a4332fbc328f9
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6072546
Commit-Queue: Chan Li <chanli@chromium.org>
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>