[Buildbucket] Add a cron job to scan builder queues
The cron job scans all of the builder Queues and check the pending and triggered builds' status. If any of the builds is ended, enqueue a PopPendingBuildTask for the build.

This is a supplemental work flow to find stuck builder queues and unstuck them.

Bug:  379656381
Change-Id: I6d7e75bbfc9a100318d719713e0526d3361b8714
Reviewed-on: https://chromium-review.googlesource.com/c/infra/luci/luci-go/+/6081931
Reviewed-by: Vadim Shtayura <vadimsh@chromium.org>
Commit-Queue: Chan Li <chanli@chromium.org>